Implemented separation of concerns into packages
